Use base64 from coreutils for encoding and decoding
The data transferred to the remote servers is encoded using base64.
This commit changes the tool used to encode and decode the data from
openssl to the base64 utility.

The change was done because openssl may be missing on some systems while
base64 should always be installed as it is part coreutils.  Moreover,
openssl needs to access openssl.cnf, which is usually located somewhere
under /etc.  This part of the file system may not be accessible to the
user on some remote machines, so openssl could fail in these cases.